+++ cmd/postgres_exporter/main.go
@@ -115,7 +115,11 @@ func main() {
 
 	prometheus.MustRegister(exporter)
 
-	pe, err := collector.NewPostgresCollector(logger, dsn)
+	pe, err := collector.NewPostgresCollector(
+		logger,
+		dsn,
+		[]string{},
+	)
 	if err != nil {
 		level.Error(logger).Log("msg", "Failed to create PostgresCollector", "err", err.Error())
 		os.Exit(1)